add task-7/1

+++ Task-3/task3_4/exceptions/ui/InvalidMenuChoiceException.java
@@ -1,7 +1,7 @@
 package task3_4.exceptions.ui;
 
 public class InvalidMenuChoiceException extends UserInputException {
-    public InvalidMenuChoiceException(int choice) {
+    public InvalidMenuChoiceException(long choice) {
         super("Пункт меню " + choice + " недоступен.");
     }
 }
